Added RangeSelectorComposition.
